backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.1
gunicorn==22.0.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mutagen==1.47.0
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
